// web/src/generated/models/NodeSource.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Revornix Main Backend
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.5.1
 * Contact: 1142704468@qq.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface NodeSource
 */
export interface NodeSource {
    /**
     * 
     * @type {number}
     * @memberof NodeSource
     */
    doc_id: number;
    /**
     * 
     * @type {string}
     * @memberof NodeSource
     */
    doc_title?: string | null;
    /**
     * 
     * @type {string}
     * @memberof NodeSource
     */
    chunk_id?: string | null;
}

/**
 * Check if a given object implements the NodeSource interface.
 */
export function instanceOfNodeSource(value: object): value is NodeSource {
    if (!('doc_id' in value) || value['doc_id'] === undefined) return false;
    return true;
}

export function NodeSourceFromJSON(json: any): NodeSource {
    return NodeSourceFromJSONTyped(json, false);
}

export function NodeSourceFromJSONTyped(json: any, ignoreDiscriminator: boolean): NodeSource {
    if (json == null) {
        return json;
    }
    return {
        
        'doc_id': json['doc_id'],
        'doc_title': json['doc_title'] == null ? undefined : json['doc_title'],
        'chunk_id': json['chunk_id'] == null ? undefined : json['chunk_id'],
    };
}

export function NodeSourceToJSON(json: any): NodeSource {
    return NodeSourceToJSONTyped(json, false);
}

export function NodeSourceToJSONTyped(value?: NodeSource | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'doc_id': value['doc_id'],
        'doc_title': value['doc_title'],
        'chunk_id': value['chunk_id'],
    };
}